requirements.txt
flask>=2.0.0
pillow>=9.0.0
pymupdf>=1.22.0
gunicorn>=21.0.0
